fix: refactored StylingComponents folder and corresponding files
