+++ frontend/src/components/AdminComponents/EditGame.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import Header from './Stylings/Header';
+import Header from '../StylingComponents/Header';
 
 const Game = () => {
