+++ frontend/src/components/AuthComponents/Register.js
@@ -4,8 +4,7 @@ import { Link } from "react-router-dom";
 import { storeUserDB } from "../../utils/dbstorage";
 import firebase from "../../utils/firebase";
 import Form from 'react-bootstrap/Form'
-import Header from "../Stylings/Header";
-import '../Stylings/Components.css';
+import Header from "../StylingComponents/Header";
 import { AuthContext } from "../../utils/Auth";
 
 const Register = ({ history }) => {
