+++ frontend/src/components/GameComponents/CurrentGames.js
@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from "react";
 import firebase from "../../utils/firebase";
 import { getPlayerInfo } from "../../utils/gamedbstorage";
-import { storePlayer } from "../../utils/localstorage";
 import PhoneLogin from "../AuthComponents/PhoneLogin";
-import './Stylings/Components.css';
-import Header from './Stylings/Header';
-import NavBar from "./Stylings/NavBar";
+import Header from '../StylingComponents/Header';
+import NavBar from "../StylingComponents/NavBar";
 
 const CurrentGames = (props) => {
