+++ frontend/src/components/HomeComponents/Home.js
@@ -1,10 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import firebase from "../../utils/firebase";
-import { clearUser, getUser } from "../../utils/localstorage.js";
-import './Stylings/Components.css';
-import Header from './Stylings/Header';
-import NavBar from "./Stylings/NavBar";
+import { getUser } from "../../utils/localstorage.js";
+import Header from '../StylingComponents/Header';
+import NavBar from "../StylingComponents/NavBar";
 
 const Home = () => {
