+++ frontend/src/components/SquadComponents/SquadCreate.js
@@ -1,7 +1,7 @@
 import React, { Component, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import Form from 'react-bootstrap/Form'
-import Header from './Stylings/Header';
+import Header from '../StylingComponents/Header';
 import { storeSquadDB } from '../../utils/squadstorage';
 
 const SquadCreate = ({history}) => {
