+++ frontend/src/components/StylingComponents/NavBar.js
